+++ src/RuleSet/Php56.php
@@ -869,7 +869,7 @@ final class Php56 implements ExplicitRuleSet, RuleSet
         return Rules::fromArray($this->rules);
     }
 
-    public function targetPhpVersion(): PhpVersion
+    public function phpVersion(): PhpVersion
     {
         return $this->targetPhpVersion;
     }