+++ src/RuleSet/Php70.php
@@ -870,7 +870,7 @@ final class Php70 implements ExplicitRuleSet, RuleSet
         return Rules::fromArray($this->rules);
     }
 
-    public function targetPhpVersion(): PhpVersion
+    public function phpVersion(): PhpVersion
     {
         return $this->targetPhpVersion;
     }