+++ test/Unit/RuleSet/AbstractRuleSetTestCase.php
@@ -31,7 +31,7 @@ abstract class AbstractRuleSetTestCase extends Framework\TestCase
         self::assertEquals($this->expectedCustomFixers(), $ruleSet->customFixers());
         self::assertEquals($this->expectedName(), $ruleSet->name());
         self::assertEquals($this->expectedRules(), $ruleSet->rules());
-        self::assertEquals($this->expectedTargetPhpVersion(), $ruleSet->targetPhpVersion());
+        self::assertEquals($this->expectedPhpVersion(), $ruleSet->phpVersion());
     }
 
     final public function testRuleSetDoesNotConfigureRulesThatAreNotRegistered(): void
@@ -222,7 +222,7 @@ abstract class AbstractRuleSetTestCase extends Framework\TestCase
 
     abstract protected function expectedRules(): Rules;
 
-    abstract protected function expectedTargetPhpVersion(): PhpVersion;
+    abstract protected function expectedPhpVersion(): PhpVersion;
 
     /**
      * @throws \RuntimeException