+++ test/Unit/RuleSet/Php53Test.php
@@ -843,7 +843,7 @@ final class Php53Test extends ExplicitRuleSetTestCase
         ]);
     }
 
-    protected function expectedTargetPhpVersion(): PhpVersion
+    protected function expectedPhpVersion(): PhpVersion
     {
         return PhpVersion::create(
             PhpVersion\Major::fromInt(5),