+++ test/Unit/RuleSet/Php55Test.php
@@ -854,7 +854,7 @@ final class Php55Test extends ExplicitRuleSetTestCase
         ]);
     }
 
-    protected function expectedTargetPhpVersion(): PhpVersion
+    protected function expectedPhpVersion(): PhpVersion
     {
         return PhpVersion::create(
             PhpVersion\Major::fromInt(5),