+++ test/Unit/RuleSet/Php73Test.php
@@ -858,7 +858,7 @@ final class Php73Test extends ExplicitRuleSetTestCase
         ]);
     }
 
-    protected function expectedTargetPhpVersion(): PhpVersion
+    protected function expectedPhpVersion(): PhpVersion
     {
         return PhpVersion::create(
             PhpVersion\Major::fromInt(7),